Generate output manifest.yaml
This makes us generate a copy of manifest.yaml containing only the
images we actually generated and with the calculated filename so
that it's easier to read in a script.

Note that this changes the generated_by command list to a list of
commands instead of a multi-line string (which was silly to begin
with) for better consistency.

Change-Id: Ic21edbc305cc6ad670abe25dbd077397f87ab451